requirements.txt
azure.eventhub==5.7.0
requests==2.26.0